fix: github dora config api (#8032)
